pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-platform"
version = "1.0.0"
description = "Enterprise AI RAG Platform — OpenRAG + RAGFlow + OpenClaw"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "python-multipart==0.0.12",

    # Validation & settings
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",

    # LLM & RAG
    "langchain==0.3.0",
    "langchain-openai==0.2.0",
    "langchain-anthropic==0.2.0",
    "langchain-community==0.3.0",
    "langgraph==0.2.28",
    "langfuse==2.50.0",

    # Vector DB
    "qdrant-client==1.13.3",

    # Embeddings
    "sentence-transformers==3.1.1",

    # Document parsing
    "pypdf==4.3.1",
    "python-docx==1.1.2",
    "openpyxl==3.1.5",
    "python-magic==0.4.27",

    # Database
    "asyncpg==0.29.0",
    "sqlalchemy[asyncio]==2.0.35",
    "alembic==1.13.3",

    # Cache
    "redis[hiredis]==5.1.0",

    # HTTP client
    "httpx==0.27.2",

    # Auth
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",

    # Observability
    "structlog==24.4.0",
    "prometheus-client==0.21.0",
    "opentelemetry-sdk==1.27.0",
    "opentelemetry-instrumentation-fastapi==0.48b0",

    # Evaluation
    "ragas==0.1.21",

    # Utilities
    "tenacity==9.0.0",
    "orjson==3.10.7",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "httpx==0.27.2",
    "factory-boy==3.3.1",
    "ruff==0.6.8",
    "mypy==1.11.2",
    "pre-commit==3.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "ANN"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
